Fix #6
This issue is due to a combination of a multiline entry, and another
value for the same postal code that isn't part of the multiline entry.
This only happens for one postal code and caused issues with the romaji
setting code.

Fixing this properly is somewhat complicated because it will require
tracking multi-line state, which the romaji code avoided doing. So this
fix is just a check for the specific code that's causing issues.